move context out of `discord_objects`
